+++ core/comps/ui/drtv-input.js
@@ -1,18 +1,104 @@
 'use strict';
 
-angular.module('cmUi').directive('input', [
+/**
+ * @ngdoc directive
+ * @name cmUi.directive:input
+ * @description
+ * a drtv for all input-tags
+ * every input with attr name and ng-model get registred at cmPristine service
+ *
+ * @restrict A
+ * @requires cmPristine
+ * @requires $rootScope
+ * @requires $timeout
+ *
+ * @example
+ <example module="cmDemo">
+ <file name="index.html">
+ <article class="content">
+     <div class="cm-input-ctn">
+        <input type="text" value="whoop whoop" />
+     </div>
+ </article>
+
+ <article class="content">
+     <div class="cm-input-ctn with-inside-icon">
+         <input type="text" value="whoop whoop" />
+         <i class="fa cm-write"></i>
+     </div>
+ </article>
+
+ <article class="content">
+     <div class="cm-input-ctn with-inside-icons">
+         <input type="text" value="whoop whoop" />
+         <i class="fa cm-rhino-bubble-glyph"></i>
+         <i class="fa cm-write"></i>
+     </div>
+ </article>
+
+ <article class="content">
+     <div class="cm-input-ctn with-inside-left-icon">
+         <i class="fa cm-checker"></i>
+         <input type="text" value="whoop whoop" />
+     </div>
+ </article>
+
+ <article class="content">
+     <div class="cm-input-ctn with-inside-left-icons">
+         <i class="fa cm-checker"></i>
+         <i class="fa cm-checkbox-wrong"></i>
+         <input type="text" value="whoop whoop" />
+     </div>
+ </article>
+
+ <article class="content">
+     <div class="cm-input-ctn with-inside-left-icon with-inside-icon">
+         <i class="fa cm-checker"></i>
+         <input type="text" value="whoop whoop" />
+         <i class="fa cm-checkbox-wrong"></i>
+     </div>
+ </article>
+
+ <article class="content">
+     <div class="cm-input-ctn with-inside-left-icons with-inside-icons">
+         <i class="fa cm-checker"></i>
+         <i class="fa cm-checker"></i>
+         <input type="text" value="whoop whoop" />
+         <i class="fa cm-checkbox-wrong"></i>
+         <i class="fa cm-checkbox-wrong"></i>
+     </div>
+ </article>
+
+ </file>
+ </example>
+ */
+
+angular.module('cmUi')
+.directive('input', [
     'cmPristine',
-    '$rootScope',
-    '$timeout',
-    function (cmPristine, $rootScope, $timeout) {
+    '$rootScope', '$timeout',
+    function (cmPristine,
+              $rootScope, $timeout) {
         return {
             restrict: 'EA',
             require: '?ngModel',
             link: function (scope, element, attrs, ngModel) {
-                if (!ngModel) return; // do nothing if no ng-model #https://docs.angularjs.org/api/ng/type/ngModel.NgModelController
+                if (!ngModel || 'cmPristineIgnore' in attrs || 'type' in attrs && attrs.type == 'hidden')
+                    return; // do nothing if no ng-model
 
-                var initValue = '',
-                    timeout;
+                // old simple style
+                //function callback_pristine(){
+                //    $rootScope.$broadcast('pristine:false');
+                //    $rootScope.$apply();
+                //}
+                //element.on('keydown',callback_pristine);
+                //
+                //scope.$on('$destroy',function(){
+                //    element.off('click',callback_pristine);
+                //});
+
+                var timeout,
+                    initValue;
 
                 function broadcastPristine(bool){
                     if(bool){
@@ -31,18 +117,22 @@ angular.module('cmUi').directive('input', [
                     ngModel.$commitViewValue();
                 }
 
-                function handleChange(){
-                    if(timeout){
-                        $timeout.cancel(timeout);
+                function handleChange(event){
+                    // check defaultValue
+                    if(initValue == undefined) {
+                        reinit();
                     }
 
+                    // call service if input is pristine
                     if(initValue != getValue()){
                         broadcastPristine();
                     } else if(initValue == getValue()){
                         broadcastPristine(true);
                     }
 
+                    // handle adaptive change
                     if('cmAdaptiveChange' in attrs){
+                        $timeout.cancel(timeout);
                         timeout = $timeout(function(){
                             setValue()
                         },attrs.cmAdaptiveChange || 1000);
@@ -51,21 +141,22 @@ angular.module('cmUi').directive('input', [
                     }
                 }
 
-                function init(){
+                // register at service
+                cmPristine.add(ngModel);
+
+                // on init or after submit
+                function reinit(){
                     initValue = getValue();
-                    cmPristine.add(ngModel)
                 }
+                cmPristine.on('reinit', reinit);
 
-                init();
-
+                // watch on element
                 element
                     .unbind('input')
                     .unbind('keydown')
                     .on('focus', handleChange)
-                    //.on('keydown', handleChange)
                     .on('keyup', handleChange)
                     .on('blur', handleChange);
-
 
                 scope.$on('$destroy', function(){
                     element
@@ -74,7 +165,9 @@ angular.module('cmUi').directive('input', [
                         .off('blur', handleChange);
 
                     cmPristine.remove(ngModel);
-                })
+
+                    cmPristine.off('reinit', reinit);
+                });
             }
         }
     }