+++ resource/phonegap/config.js
@@ -5,7 +5,7 @@ var phonegapCameoConfig = {
 // device plugin handle that
 function deviceReady(){
     document.addEventListener('deviceready', function () {
-        console.log('deviceready triggered and now phonegapCameoConfig.deviceReady=true')
+        //console.log('deviceready triggered and now phonegapCameoConfig.deviceReady=true')
         phonegapCameoConfig.deviceReady = true;
     });
 }